functions/UpdateContentFunction/requirements.txt
azure-functions
pyodbc
requests 